Included error validations for several cases

+++ app.js
@@ -5,7 +5,11 @@ const mongoose = require('mongoose');
 const path=require("path");
 var methodOverride = require('method-override');
 let  ejsMate= require('ejs-mate');
-
+let wrapAsync=require("../wanderlust/utils/wrapAsync");
+let ExpressError=require("../wanderlust/utils/expresserror.js");
+const Joi = require('joi');
+const validationSchema=require("./schema.js");
+const Schema = require("./schema.js");
 
 app.set("view engine", "ejs");
 app.set("views", path.join(__dirname, "/views"));
@@ -14,10 +18,22 @@ app.use(methodOverride('_method'));
 app.engine('ejs', ejsMate);
 app.use(express.static(path.join(__dirname, "/public")));
 
+
 main().catch(err => console.log(err));
 
 async function main() {
   await mongoose.connect('mongodb://127.0.0.1:27017/wanderlust');
+}
+
+function checkValidation(req,res,next){
+  let result=Schema.validate(req.body);
+  console.log(result);
+  if(result.error){
+    throw new ExpressError(400, result.error.message);
+  }else{
+    next();
+  }
+
 }
 
 app.listen("8080", (req,res)=>{
@@ -36,54 +52,71 @@ app.get("/listing/new", (req,res)=>{
     res.render("listings/add.ejs");
 });
 
-app.post("/listing", (req,res)=>{
+app.post("/listing",checkValidation ,wrapAsync( async(req,res,next)=>{
+ try{
   let data=req.body;
   let newListing=new Listing(data);
-  newListing.save();
+  await newListing.save();
   res.redirect("/");
-});
+ }catch(err){
+  next(err);
+ }
+}));
 
 //detail show
-app.get("/listing/:id", async(req,res)=>{
+app.get("/listing/:id",wrapAsync( async(req,res)=>{
     let {id}=req.params;
     let data=await Listing.findById(id);
+    if(!data){
+      throw new ExpressError(400, "Invalid Id!");
+    }
     res.render("listings/detail.ejs", {data});
-});
+}));
 
 //update 
 app.get("/listing/:id/edit", async(req,res)=>{
     let {id}=req.params;
     let data=await Listing.findById(id);
-    console.log(data);
     res.render("listings/update.ejs", {data});
 });
 
-app.patch("/listing/:id", async(req,res)=>{
-   let {id}=req.params;
-   function removeEmptyStrings(obj) {
-    // Create a new object to store the result
-    const newObj = {};
-  
-    // Loop through each key in the object
-    for (const key in obj) {
-      // Check if the value is not an empty string
-      if (obj[key] !== "") {
-        // Add the key-value pair to the new object
-        newObj[key] = obj[key];
-      }
-    }
-  
-    return newObj;
-  }
-
-  const result = removeEmptyStrings(req.body);
-  console.log(result); 
-   await Listing.updateOne({"_id": id}, result);
-   res.redirect(`/listing/${id}`);
-});
+app.patch("/listing/:id",checkValidation ,wrapAsync( async(req,res,next)=>{
+  let {id}=req.params;
+  function removeEmptyStrings(obj) {
+   // Create a new object to store the result
+   const newObj = {};
+ 
+   // Loop through each key in the object
+   for (const key in obj) {
+     // Check if the value is not an empty string
+     if (obj[key] !== "") {
+       // Add the key-value pair to the new object
+       newObj[key] = obj[key];
+     }
+   }
+ 
+   return newObj;
+ }
+if(!req.body){
+  throw new ExpressError(400, "Invalid Data!");
+}
+ const result = removeEmptyStrings(req.body);
+ console.log(result); 
+  await Listing.updateOne({"_id": id}, result);
+  res.redirect(`/listing/${id}`);
+}));
 
 app.delete("/listing/:id", async(req,res)=>{
   let {id}= req.params;
   await Listing.deleteOne({"_id": id});
   res.redirect("/");
+});
+
+app.all("*", (req,res,next)=>{
+  next(new ExpressError(404, "Page not found!"));
+});
+
+app.use((err,req,res,next)=>{
+  let {statusCode=500, message="Something is wrong!"}=err;
+  res.status(statusCode).render("listings/error.ejs", {message});
 });
